backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.6.1
google-generativeai>=0.4.0
httpx>=0.26.0
